style: Move user-id.decorator to users module
